item view hooked up with search, edit is left
